# cmd/api/docs/swagger.yaml
definitions:
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Zhooze is an E-commerce platform to purchase and sell shoes
  title: Go + Gin Zhooze E-Commerce API
  version: 1.0.0
paths:
...
  /user/wishlist:
    delete:
      consumes:
      - application/json
      description: Delete From wish List
      parameters:
      - description: product id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete From Wishlist
      tags:
      - WishList Management
    get:
      consumes:
      - application/json
      description: Display wish List
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Display Wishlist
      tags:
      - WishList Management
    post:
      consumes:
      - application/json
      description: Add To wish List
      parameters:
      - description: product id
        in: query
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add to Wishlist
      tags:
      - WishList Management
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"